return a userid instead of the Twitter::User object
as per the documentation - follows should be a comma separated list of User IDs, however there appears to be logic that handles getting based a list of usernames instead via the `find_user` function.  However, this function returned an instance of the Twitter::User object  instead of the user_id of the user. 

This PR changes the behavior from returning a Twitter::User object to returning the user_id.